move jetColorMap in a separate file
